requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.6
iniherit==0.3.9
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pony==0.7.3
psycopg2==2.7.3.2
python-slugify==1.2.4
requests==2.18.4
six==1.11.0
toml==0.9.4
Unidecode==1.0.22
urllib3==1.22
Werkzeug==0.14.1
yoyo-migrations==5.0.5
